Skip bash script, launch jupyter directly
The previous version of this post advised to make a startup bash script that in turn launches `jupyter`. With that setup however, it is not possible to allow Jupyter to access files in e.g. Documents, Desktop and network locations. By launching Jupyter directly, you now get a popup (on recent Mac systems) asking you whether you trust `python` to access those locations.
This popup would not show otherwise, since it would have to ask whether you want to give `/bin/bash` these permissions, and Apple has wisely blocked that suggestion from the dialog.

I think launching the shell script, and the shell hooks therein, is not necessary, since all paths are otherwise known to the python executable. The shell hook is mostly relevant for interactive shells, which we are not using in this context.