Remove email address
When I wrote this post I made up a fake email off the top of my head. It turned out to be a real address and the owner contacted me on Twitter and asked me to remove it. Replacing it with a new fake email that definitely doesn't belong to anyone.